requirements.txt
# Python Dependencies für Excel → Supabase Sync
# Installation: pip3 install -r requirements.txt

# HTTP Requests für API-Aufrufe & Dropbox Downloads
requests>=2.31.0

# Excel/Pandas für Excel-Datei Verarbeitung
pandas>=2.0.0
openpyxl>=3.1.0

# Environment Variables (.env Datei)
python-dotenv>=1.0.0

# Stock Market Data (Yahoo Finance)
yfinance>=0.2.40

# Supabase (PostgreSQL)
supabase>=2.0.0

# Anthropic API (Claude Haiku for AI enrichment)
anthropic>=0.40.0
